Initial unit test fixes after testnet reset

+++ test/java/nxt/http/TestCurrencyMint.java
@@ -77,7 +77,7 @@ public class TestCurrencyMint extends BlockchainTest {
                 build();
         JSONObject getMintingTargetResponse = apiCall.invoke();
         Logger.logDebugMessage("getMintingTargetResponse: " + getMintingTargetResponse);
-        Assert.assertEquals(4000, getMintingTargetResponse.get("difficulty"));
+        Assert.assertEquals((long)4000, getMintingTargetResponse.get("difficulty"));
         Assert.assertEquals("a9f1d24d62105839b4c876be9f1a2fdd24068195438b6ce7fba9f1d24d621000", getMintingTargetResponse.get("targetBytes"));
     }
 }
